+++ api/controller/handler/like.go
@@ -6,8 +6,8 @@ import (
 	"strconv"
 
 	"github.com/gin-gonic/gin"
-	"github.com/htoyoda18/TweetAppV2/api/handler/request"
-	"github.com/htoyoda18/TweetAppV2/api/handler/response"
+	"github.com/htoyoda18/TweetAppV2/api/controller/handler/request"
+	"github.com/htoyoda18/TweetAppV2/api/controller/handler/response"
 	"github.com/htoyoda18/TweetAppV2/api/shared"
 	"github.com/htoyoda18/TweetAppV2/api/usecase"
 	"gorm.io/gorm"
