+++ api/controller/handler/user.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/htoyoda18/TweetAppV2/api/handler/request"
-	"github.com/htoyoda18/TweetAppV2/api/handler/response"
+	"github.com/htoyoda18/TweetAppV2/api/controller/handler/request"
+	"github.com/htoyoda18/TweetAppV2/api/controller/handler/response"
 	"github.com/htoyoda18/TweetAppV2/api/shared"
 	"github.com/htoyoda18/TweetAppV2/api/usecase"
